refactor: extract OpenRouter payload transport and decoding
